exporting sha3 merkle tree instantiation
